Chatbot/requirements.txt
numpy
playwright
openai
langgraph
fastapi
uvicorn
pydantic
python-multipart
pinecone
python-dotenv
python-dateutil